+++ csatar-plugins/csatar/models/Scout.php
@@ -635,6 +635,12 @@ class Scout extends OrganizationBase
             'table' => 'csatar_csatar_mandates',
             'label' => 'csatar.csatar::lang.plugin.admin.mandate.mandates',
         ],
+        'mandatesInactive' => [
+            '\Csatar\Csatar\Models\Mandate',
+            'scope' => 'inactive',
+            'table' => 'csatar_csatar_mandates',
+            'label' => 'csatar.csatar::lang.plugin.admin.mandate.mandates',
+        ],
         'membership_cards' => \Csatar\Csatar\Models\MembershipCard::class
     ];
 
@@ -1319,6 +1325,22 @@ class Scout extends OrganizationBase
 
         return $savedCountry ?? $team->district->association->country;
     }
+
+    public function deletePersonalInformation() {
+        $this->family_name = null;
+        $this->given_name = null;
+        $this->email = null;
+        $this->phone = null;
+
+        $this->address_country = null;
+        $this->address_county = null;
+        $this->address_zipcode = null;
+        $this->address_location = null;
+        $this->address_street = null;
+        $this->address_number = null;
+
+        $this->ignoreValidation = true;
+        $this->forceSave();
 
     public function getTeamChangeHistory()
     {